rename proj to elasticsearch6
rename pkg

fix dsl imports

es imports

fix

come on

more
