+++ README
@@ -104,7 +104,7 @@ Let's rewrite the example using the Python DSL:
 .. code:: python
 
     from elasticsearch import Elasticsearch
-    from elasticsearch_dsl import Search
+    from elasticsearch6_dsl import Search
 
     client = Elasticsearch()
 
@@ -145,7 +145,7 @@ Let's have a simple Python class representing an article in a blogging system:
 .. code:: python
 
     from datetime import datetime
-    from elasticsearch_dsl import DocType, Date, Integer, Keyword, Text, connections
+    from elasticsearch6_dsl import DocType, Date, Integer, Keyword, Text, connections
 
     # Define a default Elasticsearch client
     connections.create_connection(hosts=['localhost'])
