+++ elasticsearch6_dsl/exceptions.py
