+++ elasticsearch6_dsl/mapping.py
