+++ test_elasticsearch_dsl/conftest.py
@@ -2,12 +2,12 @@
 
 import os
 
-from elasticsearch.helpers import bulk
-from elasticsearch.helpers.test import SkipTest, get_test_client
+from elasticsearch6.helpers import bulk
+from elasticsearch6.helpers.test import SkipTest, get_test_client
 from mock import Mock
 from pytest import fixture, skip
 
-from elasticsearch_dsl.connections import connections
+from elasticsearch6_dsl.connections import connections
 from .test_integration.test_data import DATA, FLAT_DATA, create_git_index, \
     create_flat_git_index
 
@@ -118,7 +118,7 @@ def dummy_response():
 
 @fixture
 def aggs_search():
-    from elasticsearch_dsl import Search
+    from elasticsearch6_dsl import Search
     s = Search(index='flat-git')
     s.aggs\
         .bucket('popular_files', 'terms', field='files', size=2)\
