+++ test_elasticsearch_dsl/test_query.py
@@ -1,4 +1,4 @@
-from elasticsearch_dsl import query, function
+from elasticsearch6_dsl import query, function
 
 from pytest import raises
