+++ test_elasticsearch_dsl/test_utils.py
@@ -1,6 +1,6 @@
 import pickle
 
-from elasticsearch_dsl import utils, serializer
+from elasticsearch6_dsl import utils, serializer
 
 def test_attrdict_pickle():
     ad = utils.AttrDict({})
