+++ test_elasticsearch_dsl/test_validation.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
-from elasticsearch_dsl import DocType, Nested, Text, Date, Object, Boolean, Integer, InnerDoc
-from elasticsearch_dsl.exceptions import ValidationException
+from elasticsearch6_dsl import DocType, Nested, Text, Date, Object, Boolean, Integer, InnerDoc
+from elasticsearch6_dsl.exceptions import ValidationException
 
 from pytest import raises
